fix scroll restoration
