+++ videoTag.client/src/main.tsx
@@ -3,9 +3,12 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
+
 import { Notifications } from "notifications.tsx";
-import { Root } from "routes/root";
+import { Home } from "routes/home";
+import { Root } from 'routes/root';
 import { Video } from "routes/video";
+
 import './main.css';
 
 const queryClient = new QueryClient({
@@ -20,11 +23,11 @@ const queryClient = new QueryClient({
 const router = createBrowserRouter([
   {
     path: '/',
-    element: <Root />
-  },
-  {
-    path: 'videos/:videoId',
-    element: <Video />,
+    element: <Root />,
+    children: [
+      { index: true, element: <Home /> },
+      { path: 'videos/:videoId', element: <Video /> }
+    ],
   },
 ]);
