--- videoTag.client/src/routes/root/constants.ts
@@ -1,13 +1,0 @@
-export const SortBy = {
-  LastModified: 'lastModified',
-  Title: 'title',
-  Size: 'size',
-} as const;
-
-export type SortByType = typeof SortBy[keyof typeof SortBy];
-
-export const QueryParam = {
-  EditMode: 'editMode',
-  TagIds: 'tagIds',
-  SortBy: 'sortBy',
-} as const;
