+++ videoTag.client/src/routes/root/index.ts
@@ -1,1 +1,1 @@
-export * from './root.tsx';
+export * from './root';
